Improve function definition
