+++ itfit/fit_functions/__init__.py
@@ -19,11 +19,18 @@ except NameError:
     __FITTER_FIT_FUNCTIONS_IMPORTED__= False
 
 if not __FITTER_FIT_FUNCTIONS_IMPORTED__:
-    from .generic_fitter import GenericFitter, GenericFitterTool
+    from .generic.generic_fitter import GenericFitter, GenericFitterTool
+    from .generic.function_container import FunctionContainer
     from .linear import LineTool
     from .quadratic import QuadraticTool
     from .exponential import ExponentialTool
     from .gaussian import GaussianTool
     from .lorentzian import LorentzianTool
+    
+    Line = FunctionContainer(linear.LineFitter, None)
+    Quadratic = FunctionContainer(quadratic.QuadraticFitter, None)
+    Exponential = FunctionContainer(exponential.ExponentialFitter, None)
+    Gaussian = FunctionContainer(gaussian.GaussianFitter, None)
+    Lorentzian = FunctionContainer(lorentzian.LorentzianFitter, None)
     
 __FITTER_FIT_FUNCTIONS_IMPORTED__ = True
